pyproject.toml
[tool.poetry]
name = "vector-app"
version = "0.1.0"
description = "A Python app with Milvus and OpenAI integration"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
pymilvus = {extras = ["model"], version = "^2.4.8"}
ell-ai = "^0.0.14"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
